+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # SwiftSQL 0.x
 
+## SwiftSQL 0.3.0
+Author: Jason Jobe
+*Dec 3, 2021
+
+- `SQLRow.columnNames` added
+
 ## SwiftSQL 0.2.0
 
 *May 14, 2020*